# specs/001-agentic-sdlc-url-shortener/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agentic SDLC URL Shortener API
  version: 3.0.0
  description: >
    Application-plane API for the URL shortener, plus read-only control-plane inspection.
    Versioning: additive changes are MINOR; removing or narrowing a field, renaming, or changing a
    status code is MAJOR and requires a new path prefix. See plan.md section 2 for the full
    compatibility and change-approval rules.

    Redirect resolution is public and MUST NOT require authentication (FR-URL-018).
    Link creation and analytics require a creator API key.

    Control-plane operations (run creation, run inspection, gate decisions) carry NO credential. The shortener's
    actor model and the orchestrator's actor model are separate domains and no credential class crosses the
    boundary (CN-012, owner ruling 2026-09-20). Orchestrator governance surfaces sit under the machine-access
    trust boundary, as creator provisioning does (FR-URL-019).

    Redirects are TEMPORARY, never permanent — resolved at the Gate 4 closing package (DF-002, CR-003).
    A permanent redirect would be cached by clients, after which follows never reach the service, defeating
    both analytics counting (FR-URL-010) and expiry enforcement (FR-URL-008). The exact code within the
    temporary class is an implementation detail (CN-006).
servers:
  - url: http://localhost:8080
    description: Local demonstration instance
...
paths:
...
  /v1/runs/{runId}/gates/{gateId}/decision:
    post:
      tags: [orchestration]
      summary: Record a human decision on a pending gate
      operationId: recordGateDecision
      description: >
        Records one of the four submittable outcomes against a pending gate (FR-ORC-013).

        TIMED_OUT is deliberately NOT submittable: it is produced by the gate-wait deadline expiring, never by a
        caller, and silence is never an input. Absence of a call to this operation is what produces suspension.

        repositoryRecordPath is REQUIRED: the decision's repository record must already exist in the working tree
        before the decision takes effect (Constitution v1.1.0 Gate semantics, FR-ORC-013). The contract refuses a
        decision that exists only in workflow state.

        Carries NO credential (CN-012). Actor authority is asserted from the recorded actor, and an agent actor is
        refused for any approving outcome at every autonomy setting (FR-ORC-021, EC-024).
      security: []
      parameters:
        - name: runId
          in: path
          required: true
          schema: { type: string }
        - name: gateId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GateDecisionRequest' }
      responses:
        '200':
          description: Decision recorded; the run state reflects its effect
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunInspection' }
              examples:
                decisionRecorded:
                  value:
                    runId: "0f7d2c18-6a4e-4f8b-9d3a-1c2b3e4f5a60"
                    state: "RUNNING"
                    terminalState: null
                    policySetVersion: "policy-set-1.1.0"
                    suspensionReason: null
                    autoAbandonAt: null
                    blockingCondition: null
                    nodes:
                      - nodeKey: "S4"
                        stageNumber: 4
                        nodeRole: "SINGLETON"
                        name: "Human clarification"
                        state: "SUCCEEDED"
                        executorClass: "HUMAN_GATE"
                        executorKindUsed: "HUMAN"
                        attemptsUsed: 1
        '400':
          description: >
            Malformed decision — unknown outcome, CHANGES_REQUESTED with an empty change list, or a missing
            repositoryRecordPath.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                missingRepositoryRecord:
                  value:
                    code: "INVALID_INPUT"
                    message: "repositoryRecordPath is required and must begin with docs/governance/."
                    detail: "A gate decision is not usable until its repository record exists (CR-013)."
        '403':
          description: >
            The actor has no recorded authority for this gate class, or an agent actor attempted an approving
            outcome. Recorded as an autonomy violation (EC-024, FR-ORC-021).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                nonHumanActorRefused:
                  value:
                    code: "FORBIDDEN"
                    message: "Only a human actor may record a gate decision."
                    detail: "actorType accepts only human (CR-021). Actor identity is declared, not verified:see FR-ORC-021 for the disclosed limitation."
        '404':
          description: No such run, or no such pending gate on it
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                gateNotFound:
                  value:
                    code: "NOT_FOUND"
                    message: "No open gate with that identifier on this run."
        '409':
          description: >
            The gate already carries a decision. Gate decisions are append-only and corrected by a superseding
            record, never overwritten (Compensation Register).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                gateAlreadyDecided:
                  value:
                    code: "CONFLICT"
                    message: "That gate already carries a recorded decision."
                    detail: "A decision is not replaceable; a changed decision is a new record referencing theprior one."
...
components:
  securitySchemes:
    creatorApiKey:
      type: http
      scheme: bearer
      description: >
        Creator credential presented as `Authorization: Bearer <key>`.

        Standard bearer transport is chosen deliberately: proxies, frameworks and log scrubbers redact the
        Authorization header by default, whereas a custom header receives no such default protection. This is
        free reduction of exactly the leak risk FR-URL-017 treats as non-waivable.

        Key form: `crk_` + base64url of at least 256 CSPRNG bits. The prefix makes a leaked key identifiable on
        sight and gives the POL-SEC-002 secret scan a deterministic pattern to match; it carries no information
        and does not reduce entropy. The stored form is SHA-256 of the full presented string, prefix included.

        Provisioned by the local operator script only; there is no HTTP issuance endpoint (FR-URL-019). The
        script requires an expiry parameter — a duration, or the explicit literal `never`; there is no default.
        Never logged by the application at any level.

        Expired and revoked credentials are refused with an identical response shape, so a caller learns
        nothing from the distinction.
  schemas:
...
    GateDecisionRequest:
      type: object
      required: [outcome, actorType, actorName, reason, repositoryRecordPath]
      additionalProperties: false
      properties:
        outcome:
          type: string
          enum: [APPROVED, REJECTED, CHANGES_REQUESTED, ESCALATED]
          description: >
            Four submittable outcomes. TIMED_OUT is absent by design — it is produced by deadline expiry, never
            submitted, because silence must never be expressible as an input (Constitution III).
        actorType:
          type: string
          enum: [human]
          description: >
            Only a human may submit a gate decision. The system actor is permitted solely for deadline expiry and
            retention-driven abandonment, neither of which is an approval and neither of which uses this operation.
        actorName: { type: string, minLength: 1 }
        reason:
          type: string
          minLength: 1
          description: Mandatory. A decision without a reason is not a gate record.
        repositoryRecordPath:
          type: string
          pattern: '^docs/governance/'
          description: >
            Path to the decision's materialized repository record, which MUST already exist. Writing that record is
            not itself work the decision authorizes, so this rule cannot deadlock (FR-ORC-013).
        changes:
          type: array
          items: { type: string, minLength: 1 }
          description: Required and non-empty when outcome is CHANGES_REQUESTED; the changes are enumerated, never implied.
        escalationTarget:
          type: string
          minLength: 1
          description: >
            Required when outcome is ESCALATED: what decision exceeds the current owner's authority (CR-041).
            A recorded string, declared by the submitting actor the same way actorName is — there is no user
            registry this is verified against. The system records and surfaces it (run inspection, the audit
            trail) but does not notify or deliver it to anyone; a production deployment would wire a notifier
            on top of this field, which this system does not do.
        noPlanOption:
          type: string
          enum: [GOVERNANCE_ONLY, HUMAN_IMPLEMENTED, ABANDON]
          description: >
            Required only on the no-change-plan gate class (FR-ORC-031, CR-001). HUMAN_IMPLEMENTED records
            executorKindUsed HUMAN on the implementation node.
      allOf:
        - if: { properties: { outcome: { const: ESCALATED } } }
          then: { required: [escalationTarget] }
...
    RunInspection:
      type: object
      required: [runId, state, policySetVersion, nodes]
      additionalProperties: false
      properties:
        runId: { type: string }
        state:
          type: string
          enum: [PENDING, RUNNING, SAFE_STOP, COMPLETED, REJECTED, ABANDONED]
        terminalState:
          type: [string, 'null']
          enum: [COMPLETED, REJECTED, ABANDONED, null]
          description: Null while suspended — a suspended run has no terminal outcome (CL-005).
        policySetVersion: { type: string }
        suspensionReason: { type: [string, 'null'] }
        autoAbandonAt:
          type: [string, 'null']
          format: date-time
          description: Non-null iff state is SAFE_STOP.
        blockingCondition: { type: [string, 'null'] }
        nodes:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [nodeKey, stageNumber, nodeRole, name, state, executorClass]
            properties:
              nodeKey:
                type: string
                pattern: '^S([1-9]|1[0-2])(\.([1-9][0-9]*|join))?$'
              stageNumber: { type: integer, minimum: 1, maximum: 12 }
              nodeRole: { type: string, enum: [SINGLETON, FAN_OUT_PARENT, FAN_OUT_CHILD, JOIN] }
              parentNodeKey: { type: [string, 'null'] }
              name: { type: string }
              state:
                type: string
                enum: [BLOCKED, READY, RUNNING, AWAITING_APPROVAL, RETRY_WAIT, FALLBACK,
                       ROLLING_BACK, COMPENSATING, SUCCEEDED, FAILED, INVALIDATED, SKIPPED]
              executorClass: { type: string, enum: [DETERMINISTIC, AI_CAPABLE, HUMAN_GATE] }
              executorKindUsed:
                type: [string, 'null']
                enum: [DETERMINISTIC, AI, HUMAN, null]
                description: >
                  Executor kind actually used. Required on every executed node (FR-ORC-029). HUMAN is recorded
                  for a node the human implemented under the no-plan gate (FR-ORC-031); it is never
                  flag-selectable.
              attemptsUsed: { type: integer, minimum: 0 }
              dependsOn:
                type: array
                items: { type: string }
                description: Node keys this node depends on, with join semantics carried on the edge (FR-ORC-003).
        pendingGate:
          type: [object, 'null']
          additionalProperties: false
          required: [gateId, nodeKey, gateClass, waitDeadline, disclosedAutoAbandonAt]
          properties:
            gateId: { type: string, description: Addresses this gate in the decision operation. }
            nodeKey: { type: string, description: The node the gate is attached to. }
            gateClass: { type: string }
            waitDeadline: { type: string, format: date-time }
            disclosedAutoAbandonAt: { type: string, format: date-time }
          description: >
            Both deadlines are disclosed at request time, so the person being asked sees the expiry
            consequence in the ask itself (CL-005 addendum).
    Error:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code: { type: string }
        message: { type: string }
        detail:
          type: string
          description: >
            Must never echo raw input in a way that enables injection, and must never contain
            secrets, credentials, or storage internals (FR-URL-002, FR-URL-017).